Reduce some members of route_t struct to uint16_t

+++ tdata.h
@@ -20,10 +20,10 @@ typedef struct route route_t;
 struct route {
     uint32_t route_stops_offset;
     uint32_t trip_ids_offset;
-    uint32_t n_stops;
-    uint32_t n_trips;
-    uint32_t attributes;
-    uint32_t operator_offset;
+    uint16_t n_stops;
+    uint16_t n_trips;
+    uint16_t attributes;
+    uint16_t operator_offset;
     rtime_t  min_time;
     rtime_t  max_time;
 };
